Move docfxBuilder to its own module file.
